set debian repository based on major version only
Only update apt-cache when needed

Instead of always updating the apt cache, leverage the 'update_cache'
method during the apt install task.  Prior to this patch, every run of
this role resulted in a "changed" state.  This patch will clean up that
report to only be "changed" when this actually did indeed change (as
updating the apt-cache isn't a meaningful "change").

use apt_repository instead of file copy